+++ lib/src/z_machine.dart
@@ -48,12 +48,12 @@ class ZMachine {
   static ZMachine? _context;
 
   //contains machine version which are supported by z-machine.
-  final List<Engine> _supportedEngines = [
-    Version3(),
-    Version4(),
-    Version5(),
-    Version7(),
-    Version8(),
+  final List<Engine Function()> _supportedEngines = [
+    () => Version3(),
+    () => Version4(),
+    () => Version5(),
+    () => Version7(),
+    () => Version8(),
   ];
 
   /// Represents the underlying interpreter engine used to run the
@@ -137,14 +137,12 @@ class ZMachine {
 
     ver = ZMachine.intToVer(rawBytes[Header.version]);
 
-    var result = _supportedEngines
-        .where(((Engine m) => m.version == ver))
-        .toList();
+    final result = _supportedEngines.where(((m) => m().version == ver)).toList();
 
     if (result.length != 1) {
       throw Exception('Z-Machine version $ver not supported.');
     } else {
-      engine = result[0];
+      engine = result[0]();
     }
 
     log.info('Zart: Using Z-Machine ${engine.version}.');
@@ -202,10 +200,7 @@ class ZMachine {
     }
 
     if (inBreak) {
-      await Z.sendIO({
-        "command": IoCommands.printDebug,
-        "message": "<<< DEBUG MODE >>>",
-      });
+      await Z.sendIO({"command": IoCommands.printDebug, "message": "<<< DEBUG MODE >>>"});
       callAsync(Debugger.startBreak);
     }
   }
@@ -220,11 +215,7 @@ class ZMachine {
     //if output stream 3 is active then we don't print,
     //Just preserve the buffer until the stream is de-selected.
     if (!engine.outputStream3) {
-      sendIO({
-        "command": IoCommands.print,
-        "window": engine.currentWindow,
-        "buffer": sbuff.toString(),
-      }).then((_) {
+      sendIO({"command": IoCommands.print, "window": engine.currentWindow, "buffer": sbuff.toString()}).then((_) {
         sbuff.clear();
       });
     }
@@ -239,6 +230,16 @@ class ZMachine {
     memoryStreams.clear();
     engine.mem = MemoryMap(rawBytes);
     engine.visitHeader();
+  }
+
+  /// Hard reset Z-Machine to initial state */
+  void hardReset() {
+    engine.programCounter = 0;
+    engine.stack.clear();
+    engine.callStack.clear();
+    memoryStreams.clear();
+    rawBytes.clear();
+    isLoaded = false;
   }
 
   void _assertLoaded() {